Make check for `+dune` accept `+dune[digit]` as well
Sometimes we need to re-release a dune port, but the existing dune port
is out there and in use, so it makes more sense to make a new release.
For this case we decided to use `+dune2`, `+dune3` etc. so this PR
extends the logic to allow digits as well.